Move tests (functions/) to jest
