+++ package.json
@@ -58,7 +58,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "test": "jest && react-scripts test --env=jsdom",
     "eject": "react-scripts eject",
     "lint": "eslint src",
     "lint-fix": "npm run lint -- --fix",
